Remove TestSetup.__parser and replace it by .__globals/.__mocks
Most of the objects stored in the CParser() referred are by
TestSetups subclass already directly (i.e. typedefs, macros, ...).
The only items not yet referred yet are functions and global variables,
which are stored in TestSetup.__globals.
Furthermore TestSetup.__mocks contains a list of all global vars/funcs
that have to be mocked.