Report C compiler errors immediately now
Up to now parsing errors of the C code were reported delayed (when
instatiating a testsetup class the first time). This caused very
unintuituve exception messages when accessing a class attribute
(i.e. a macro) of a testsetup with invalid C source code before the
first instantiation. The attribute is not available but the user is not
informed why.