Modify CRawAccess semantics
No CRawAccess has not a window into memory with start end end, but
a pointer with only a start. It MAY still have a max_size. But this
is more like an upper limit, not a guaranteed length.

Due to this modifications a comparison to a 'bytes' operand is no
done based on the length of the 'bytes' operand. This both expressions
"x == b'1'" and "x == b'12'" MAY be True.