Modify semantic of storing str/bytes object to pointer/array
Now every character of a str/bytes object is stored to one array
element of the array/pointer being assigned to. It is not possible any
more to store bytes/str objects to
array/pointers of non-scalars/non-void.

If a str has a unicode
character that does not fit into the array element (i.e. '\u1234' is
stored to pointer to uint16), it is split over multiple array elements
according to utf8 (on 8 bit array element size)/utf16
(on 16 bit array element size).