Extract AddressSpace from CDataModel
The first step of removing the ctypes dependency is to extract all
ctypes calls for accessing data objects (not function calls!)
into a separate "AddressSpace" class. The CDataModel
now handles all convertions between python- and C- data format on
its own and uses AddressSpace only for reading/writing the raw data
from/to the C memory space.

The currently used "InprocessAddressSpace" shall be replaced later by
addressspaces which are not relying on ctypes.